아이템 추가 관련 수정
아이템 클래스 수정
어댑터에서 아이템 메소드 수정

+++ app/src/main/java/k/tomorrowdecision/TextListViewAdapter.java
@@ -58,7 +58,7 @@ public class TextListViewAdapter extends BaseAdapter {
     }
 
     // 아이템 데이터 추가를 위한 함수
-    public void addItem(int id, String todo) {
+    public void addItem(long id, String todo) {
         TextItem item = new TextItem();
 
         item.setId(id);
